feat: A basic reactome inference rule based on EC-number
From the Karp et al. 2011 paper: reactions are infered to be present
based on a EC-number when it is annotated with this EC-number.

Here, we use a SPARQL endpoint as a backend knowledge base, containing
BioPAX level 3 RDF of MetaCyc.
A base test is done for EC-number 1.1.1.1, with reference to the current
state of MetaCyc dabase