feat: Reactome from EC-numbers + inverse reactome problem
A new approach to the "inverse" problem of the reactome (i.e., finding
the minimal set of monomer required to catalyze a set of reactions),
that makes it possible to solve it way faster, by reducing the set of
potential monomer to consider in the subset choice, by selecting only
those that may be involved in the catalysis.

Also, some bit of refactoring.