feat: reactome inference directly from pythoncyc
I asserted that the result reactome given by ASP based computation
is the same as the one given by querying the MetaCyc knowledge base
with PythonCyc directly on inference time, on a specific example based
on 100 randomly chosen monomer identifiers.
However, this approach is definitely slower, and requires a running
PathwayTools instance.
This can be considered rather as a validation of the ASP approach (or at
least, a sanity check).